# contracts/openapi/conexus.v1.yaml
openapi: 3.1.0
info:
  title: Conexus API
  version: 1.0.0
  description: |
    Contract for the Conexus FastAPI gateway as implemented in this repository (v0 / v1 boundary).

    **Implemented in code today:** health and readiness, OpenAI-compatible `POST /v1/chat/completions`,
    and internal adapter-profile registry routes (when enabled).

    **Not implemented on the public gateway:** `GET /v1/models`, `POST /v1/route`, `POST /v1/embeddings`,
    `GET /v1/usage/{trace_id}`, `GET /v1/traces/{trace_id}`. Those appear in the work-package as forward
    targets; see `docs/readiness/CONEXUS_APP_INTEGRATION_READINESS.md`.

    **Identifiers:** successful gateway calls return `X-Conexus-Request-Id` and a response `id` of the form
    `chatcmpl-{request_id}`. This is a per-request correlation id stored with gateway request logs; it is
    not the same as a future first-class `trace_id` field in JSON bodies.

    **Auth:** project-scoped gateway calls use `Authorization: Bearer <project_api_key>`. Internal
    adapter routes use `X-Internal-Api-Key` (and optional `X-Internal-Actor`).
  x-contract-source: merged-from-conexus-v1-contracts-work-package-2026-05-11
  x-contract-notes:
    - FastAPI auto-docs remain at /docs; this file is the portable contract for other apps.
    - JSON Schemas under contracts/json-schema/ supplement components where noted.
servers:
  - url: http://localhost:8000
    description: Local development (docker compose default)
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [Gateway]
      summary: OpenAI-compatible chat completion (minimal supported subset)
      operationId: createChatCompletion
      security:
        - ProjectApiKey: []
      parameters:
        - name: X-Conexus-Domain-Key
          in: header
          required: false
          schema:
            type: string
          description: Optional domain key for adapter-profile routing context.
        - name: X-Conexus-Gateway-Profile-Id
          in: header
          required: false
          schema:
            type: string
          description: Optional explicit gateway profile id (alias `X-Conexus-Adapter-Profile-Id`).
        - name: X-Conexus-Request-Id
          in: header
          required: false
          schema:
            type: string
            maxLength: 64
            pattern: "^[a-zA-Z0-9_-]{1,64}$"
          description: |
            Optional caller correlation id. When present and valid, Conexus stores it as the
            gateway ``request_id`` (must be unique per new request). When omitted, Conexus generates one.
            Echoed on every response in this header and, for non-streaming JSON, in ``request_id``.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionsRequest"
      responses:
        "200":
          description: Non-streaming completion, or SSE when `stream=true`
          headers:
            X-Conexus-Request-Id:
              description: Correlates with gateway request logs in the BO.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionsResponse"
            text/event-stream:
              schema:
                type: string
              description: |
                OpenAI-style SSE (`data: {json}\\n\\n`, terminal `data: [DONE]`). Error payloads in-stream
                use a documented `error` object shape (see gateway implementation).
        "400":
          description: Client / validation error from gateway domain
          headers:
            X-Conexus-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpErrorWrapper"
        "422":
          description: Request body failed FastAPI/Pydantic validation
          headers:
            X-Conexus-Request-Id:
              description: Present for POST /v1/chat/completions only; generated when body validation fails.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "429":
          description: Project limits / rate-style gateway error
          headers:
            X-Conexus-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpErrorWrapper"
        "502":
          description: Upstream provider error surfaced by Conexus
          headers:
            X-Conexus-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpErrorWrapper"
        "409":
          description: Reused X-Conexus-Request-Id for a new gateway request (unique constraint)
          headers:
            X-Conexus-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpErrorWrapper"
...
components:
  securitySchemes:
    ProjectApiKey:
      type: http
      scheme: bearer
      bearerFormat: Project API key
    InternalAdapterApiKey:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
...
  schemas:
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
    ChatCompletionsRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: Concrete provider model id or configured alias (see `backend/static_config/model_aliases.yaml`).
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatMessage"
        max_tokens:
          type: integer
          minimum: 1
          maximum: 128000
          default: 4096
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.2
        stream:
          type: boolean
          default: false
        top_p:
          type: number
          nullable: true
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          nullable: true
        user:
          type: string
          nullable: true
        response_format:
          type: object
          nullable: true
          additionalProperties: true
        seed:
          type: integer
          nullable: true
        n:
          type: integer
          nullable: true
        tools:
          type: array
          nullable: true
          items:
            type: object
            additionalProperties: true
        tool_choice:
          nullable: true
          oneOf:
            - type: object
              additionalProperties: true
            - type: string
        logprobs:
          type: boolean
          nullable: true
        top_logprobs:
          type: integer
          nullable: true
        presence_penalty:
          type: number
          nullable: true
        frequency_penalty:
          type: number
          nullable: true
    ChatChoice:
      type: object
      required: [index, message, finish_reason]
      properties:
        index:
          type: integer
        message:
          $ref: "#/components/schemas/ChatMessage"
        finish_reason:
          type: string
    TokenUsageOpenAI:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
    ChatCompletionsResponse:
      type: object
      required: [id, object, created, model, provider, fallback_used, request_id, choices, usage]
      properties:
        id:
          type: string
          example: chatcmpl-01ARANDOM
        object:
          type: string
          const: chat.completion
        created:
          type: integer
          format: int64
          description: Unix epoch seconds (integer), not RFC3339 string.
        model:
          type: string
        provider:
          type: string
        fallback_used:
          type: boolean
        request_id:
          type: string
          description: Same value as response header X-Conexus-Request-Id; gateway log row key.
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatChoice"
        usage:
          $ref: "#/components/schemas/TokenUsageOpenAI"
    ConexusErrorDetail:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        limit_type:
          type: string
          nullable: true
        current_value:
          type: integer
          nullable: true
        limit_value:
          type: integer
          nullable: true
        window:
          type: string
          nullable: true
        reset_at:
          type: string
          nullable: true
    HttpErrorWrapper:
      type: object
      properties:
        detail:
          oneOf:
            - $ref: "#/components/schemas/ConexusErrorDetail"
            - type: string
...